moved draw_loading_screen to mainmenu.c as per lao's suggestion
