+++ src/menu/mainmenu.h
@@ -14,5 +14,6 @@ void create_main_menu(MenuData *m);
 void draw_main_menu_bg(MenuData *m);
 void draw_main_menu(MenuData *m);
 void main_menu_update_spellpractice(void);
+void draw_loading_screen(void);
 
 #endif
